handler dir分割
